+++ game/ui.go
@@ -87,7 +87,7 @@ func (ui *uiManager) updateHover(s *scene, mx, my int32, inv *inventory, dt floa
 		n.itemMatch = false
 	}
 	for _, n := range s.npcs {
-		if n.silent {
+		if n.silent || n.hidden {
 			continue
 		}
 		if n.containsPoint(mx, my) {